backend/requirements-dev.txt
# Development dependencies
#
# Install production deps
-r requirements-prod.txt

# Install development deps
ipython==2.2.0
Markdown==2.5
django-filter==0.7
